Aim Introducing the new class Aim, which is part of RangedWeapon. Aim is really self-explanatory, so just look through, but notice the methods "damageAt" and "aiming" both in RangedWeapon. They will be part of new damage calculating system, which allows a damage radius with fewer damage at the borders and maximum damage in the center.

+++ src/player/weapon/AbstractArrow.java
@@ -331,7 +331,7 @@ public abstract class AbstractArrow extends RangedWeapon implements gui.Drawable
 	}
 
     @Override
-    public int getRange () {
+    public double getRange () {
         return super.getRange();
     }
